fix: slug generation
